Don't execute sudo during build.
There are systems that don't have sudo and there are people
that believe executing sudo without the user's permission is
rude. For example, a user may have run a sudo command for
something else and then the sudo in the rbx build will run
without prompting for permission.

For these reasons, it's better to just let the user handle
the situation explicitly.