Setup proper SignalException handling
What a wonderful MRI behavior we see here again. What MRI does is when a
SignalException escapes to the toplevel, it catches it, sets up the
default signal handler for that signal and reraises the signal to the
current process. We therefore mimic this behavior by setting up signal
handlers for the same signals MRI catches and raises an exception. Our
post handling then does the same by setting up the default handler and
resending the signal.

Fixes #1839