Add validation for adding global handles
We want to make sure people are only adding valid handles to the list of
global handles. This means that the VALUE pointer in the CAPI needs to
point to a proper CAPI handle and not point to some random memory
location that can cause things such as GC crashes later on.

Should help in debugging stuff like therubyracer to see earlier where
extensions such as that register invalid handles.