Cache the default_external and default_internal encoding
This happens quite a lot, especially since 27e43a67 which caused CI runs
for 1.9 mode to run at least 30 seconds longer on my system locally. We
use undefined here as the default value and as the marker to reset the
value, since nil can be a valid value here and in that case we don't
want to look it up each time either.

The ontology change was needed, because otherwise undefined was not the
correct value yet when the encodings were initialized.