Use consistent PrimitiveFailure exception messages.
Primitives can fail at any time, even during VM boot where the backtrace can
be primitive or possibly missing. The following format of the PrimitiveFailure
exception message provides all necessary information to index directly to the
Ruby method and then to the primitive implementation:

  "[module scope]::class_name[.#]method_name primitive failed"

This is the minimum required exception message. If there exists only a single
possible cause of failure, that can be appended to, but does not replace,
the basic exception message.