Introduce --libprefixdir configuration option.
This configuration option allows to specify directory, into which will
be palced lib, kernel and runtime sub-directories, which have to be in the
same folder (this precondition was broken due to fix for #1757).

It alows to restore the standard meaning of --libdir configuration
option, i.e. the standard system library directory instead of location
of Rubinius lib sub-directory.

Runinius lib, kernel and runtime subdirectory names are not configurable
ATM.