Add --build and --host configure options.
These configure options mimic standard autoconf options. The --build is
guessed by default, but it has no real meaning ATM. The --host defaults
to BUILD value, but can be overridden, to specify the platform Rubinius will
run on. This could be used for cross-compilation in the future. At now,
it helps to recognize vendor properly.