OnStack can mark a location on the stack that contains NULL
When we use OnStack, we might mark a location that holds the return
value of a call. This return value can be NULL to indicate an exception
so if we see a NULL here, that is a valid value and we just shouldn't do
anything with it.

Before this would cause a segmentation fault because we would deference
a NULL pointer.