Use a block around OnStack with setjmp / longjmp for C-API exceptions
This way we ensure that the destructor for OnStack is executed, even
when we encounter an exception. We use setjmp / longjmp in the C-API but
that means we can't depend on RAII behavior here.

By moving this into a separate block we make sure the destructor part of
RAII runs before we do the longjmp.