Changed configure --libdir to --appdir. Closes #1919.
The --libdir option is retained for specifying where to put the Rubinius
shared library. We don't build a shared or static library yet but we will
when we have an embedding API.

The --appdir option specifies where the application runtime and library
files will be installed. The use of --libdir for this was misleading.