Always emit code for pushing a block in the JIT
There are situations where we create incorrect code before this fix.

Imagine the following code:

def x
  y do
    true
  end
end

def y
  z { yield }
end

def z
  yield
end

Now if in this case we decide to compile 'x', we can inline y into x and
z into y. The problem occurs when we try to inline z but we can't due to
it being too complex, or when it blows the inline budget. In that case
the block hasn't been pushed so z will raise a NoJumpError because it
can't find the block.

In the future we need to make the JIT logic here smarter. We need to
keep track of whether we end up using the block in the whole chain down
when we compile a method and if we don't end up using the block, we can
remove the code for pushing it.

Fixes #1940