Fixed Encoding ontology.
We are going to use Encoding in 1.8 mode behind the scenes but we
cannot expose the Encoding class under Object because stupid Ruby
code in the wild looks for Encoding to decide whether to use 1.9
language syntax. So initially I moved everything under Rubinius,
but this leads to ugly eg Rubinius::EncodingClass::Encoding::XXX
output for exceptions.

This way, we only expose Encoding under Object if not in 1.8 mode.