Add caching and pre-computed Encoding::Converter paths.
It takes a fraction of a second to load a pre-computed association of all
possible conversions from the builtin converters. If other converters are
loaded, the cache will still be used, but cache misses are costly because
an exhaustive search will be performed for any missing and unavailable
conversions. However, the likelihood of those are rare, since they will
raise exceptions.

This caching reduces spec run times by about 50% for encoding specs and
about 10% overall.