Fix local storage and retrieval for inactive fibers
When a variable scope is allocated on a fiber, we need to use the
displacement so we find the proper location of the locals. When the
fiber is active, the displacement is 0, but if it's inactive, it will
not be zero. So we need to handle that case so we can properly find the
correct memory location.