Update the thread id and lock count after setting inflated header
Before we first were setting the thread id and lock count before
updating the inflated header with the current header word. This results
in actually loosing one count, because the header word still has the old
count, but the passed in count parameter is already increased by one.

This would cause an error for the last recursive unlock if we do
recursive locking so often that we need to inflate the header.