Fix object allocation race condition that could cause immix corruption
The problem here is the following case. When we had to fall back to
immix allocation because the young generation slab allocation fails, we
used the new_object_fast code path that used the allocation spinlock.

When this path was used, it almost always ends up calling
new_object_typed which also locks. Because of this there was a manual
unlock before calling new_object_typed. The bug was that because the
lock was a scope guard, we ended up unlock too often after the
new_object_fast method.

This resulted in the possibility that two threads where allocating
objects in immix which isn't thread safe at this point. This caused
memory locations to be used multiple times for different objects,
causing all kinds of weird errors.

The fix completely removes the new_object_fast code path, because unlike
the name suggests, it doesn't actually have any faster code path than
new_object_typed and ended up calling new_object_typed almost always
anyway. Therefore removing this code path is by far the simplest
solution for the problem.

We also fix inflate_for_handle so it uses the proper inflation lock
instead of using SYNC. The last change to ontology setup is more a
stylistic change because I added a bunch of assertions on object
validity which wanted an object to be valid when for example setting
instance variables.