Fix visibility lookup override
If the original method implementation had a different visibility than
the new method location we need to preserve the visibility of the new
method location.

This for example happens in you change the visibility of a method in a
singleton class. The singleton class has a new entry with a different
visibility but defers for the implementation to the original definition.

Fixes #1967