Fix initialization ordering in ontology
Encoding needs a lot of the infrastructure in place to initialize. One
important thing is that it needs structures such as CompactLookupTable
for instance variables ready. Before Encoding was initialized before
this, resulting in a CompactLookupTable that had 0x22 as it's class
pointer.

This was exposed when doing a heap dump, which then crashed because it
couldn't find any class for that invalid 0x22 pointer value. Reordering
in the ontology makes sure we setup stuff in the right order.

Fixes #1933