Introduce system threads and mark signal and finalizer thread as such
We need a concept of a system thread, which is a managed that thread
that runs Ruby code but isn't visible to userland Ruby code. Currently
that means at least the signal and the finalizer thread.

We were already hiding the signal thread, but introducing this
abstraction is a much more useful approach.