Fix type check in the JIT
We changed the object layout at some point to allow for storing C-API
handles inline. At that point, check_type_bits wasn't updated to reflect
the new object layout.

We also reorder the object header layout so we can make a check just on
the first 9 bits (1 inflation bit and 8 object type bits) so we don't
exclude valid cases where we use a non 0 meaning for the additional bits
in the header.

Fixes #1973