Made default encoding ASCII-8BIT.
The default encoding is what MRI assigns to index 0, (ie, no encoding bit
set in the object header). This implementation detail can be the cause of
obscure and latent bugs. So I chose to not use the same indexes as MRI.
However, while not changing the indexes, we must match the behavior. So,
now the default encoding is correctly set to ASCII-8BIT if no encoding
is set on a String.