Ensure that we have the machine code setup before changing executor
Previously we wrote the machine_code instance variable after setting up
the new executors. This leads to a potential race condition where
another thread already sees the new executor, but still sees a NULL
pointer for the machine.

This then crashes the VM in the execute_specialized template where we
dereference the machine code to get the number of locals for the machine
code.