Fix race condition in checks for packed ivars
Before we were reading the type information for whether it's a regular
or a packed object multiple times. Because we read it multiple times,
there's a small window that can cause a race condition.

What happens is that we read the first condition as not a packed object.
Another thread could be in the process of setting up the packing
information. That means that when we check the type again for
Object::type, the other thread has changed it already to
PackedObject::type.

This results in the code ending up in the else case where it reads the
type info again. It then allocates an object with PackedObject::type,
but without the proper size. This causes the first write to a packed
ivar for that object to corrupt memory.

By reading this type info variable only once, we make sure it doesn't
change while we check the different cases. In case of the same scenario
we now get two threads that will both try to setup the packing
information. This however is not a problem because this case is properly
guarded.