Fix mutex busy when shutting down a thread
When a thread was cleaned up, it could be that the VM lock object was
still locked in VM::wakeup. We now lock around the discard code to grab
this lock.

This means we have to change the code that releases contention, since if
it's called when the VM lock is still taken, it might deadlock.

Fixes #1952