Fix handle initialization setup
There was a race condition when we want to inflate an object header. The
problem was occured when an object was thinly locked and then an object
id needed to be stored too. In that case we need to inflate header.

During this scenario, another thread actually unlocked the thinly locked
object. This caused the atomic replacement in
ObjectHeader::set_inflated_header to fail. This is exactly how it should
work.

The problem however is that when we attempt to set the inflated header a
second time after the failed compare and swap, we don't reset the values
derived from the original flags in the header. So the owner_id_ that was
set during the first attempt, wasn't cleared on the second attempt.

This meant that after a succesful second attempt of the compare and
swap, the inflated header had an owner_id_ set, even though the second
time the original object header was not thinly locked and had no owner
anymore.

This could cause deadlocks later on, because other threads would still
see the object as locked, but since it wasn't, nothing every unlocked it
again.

This change makes sure we reset the additional properties we track in
the inflated header each time we update the inflated header.