If we try to update a header and it fails we can retry
If we fail to update a header atomically and see an empty object header,
that's a valid state. It could be that the header was used for something
else before and isn't anymore, for example when it was thinly locked
when trying to install an object id and unlocked when the header was
swapped.

Related to #1995