Don't emit code for creating a block if not necessary
When we inline blocks succesfully in the JIT, we don't need to emit code
to create this block on the stack. This is only necessary when we either
fail to inline, or when emitting the uncommon case in the jit.

This implements a better version of the behavior removed in
897570e8ced9b2c902f867316fec56837c7e7081 because that was buggy. This
implementation handles more cases and also doesn't suffer from the bug
present in that implementation.