Fix invalid memory writing in String
We were writing outside the string space in rare cases. The problem is
that when the original string had a capacity of 0 and the appended
string also has size 0. This results in a new minimal capacity of 0.

This means the string didn't need to be resized, but we still write a \0
byte at the end. This means we overwrite memory of the next object in
memory, causing corruption.

This change makes sure we always increase the capacity when needed.

Fixes #2030