Style improvements for VM code
In general we prefer using C++ style casts over C ones. Also we don't
need to cast if we already have the correct type. In a few cases we
should use our proper as<> template for casting.

In the tooling interface we also have methods for casting, so we should
use them and not do it manually.

Another thing is to make sure we initialize all properties in
constructors.

Lastly we use variable array allocation where we can instead of using
alloca() manually. This creates much simpler and easier to understand
code.