Use proper read accessor for String::encoding.
There are code locations that use a conditional that would check if the
encoding_ attribute is nil and exit early. However, those are distant from
code that does encoding()->get_encoding() so they can be missed when eg
refactoring. We should always use encoding(state) accessor unless there is
a very special purpose reason not to.