Setup new unwinds structure that allows dynamic growth
We allocate a small static space for unwinds and allow for growth
instead of large static sized area. The size of the small space is based
currently on CI runs that have enough head room with a size of 4.

This reduces the size of a call frame in the interpreter by around 180
bytes.