When closing a popen pipe, child might already have exited
In the Puma tests a server is spawned with IO.popen and in the teardown
it explicitly uses a waitpid() when shutting down. This means when
IO#close tries to do that too, it throws an error.

In this case, we can ignore the error if the child has already
disappeared.