Fix signing issue with Fixnums in the JIT
Previously we were using LLVM's lshr instruction, which is defined as:

The 'lshr' instruction (logical shift right) returns the first operand
shifted to the right a specified number of bits with zero fill.

This means it doesn't consider the sign bit. This means that parts in
the JIT that were using this, could cause wrong numbers to be created if
a negative fixnum was shifted so it could be used as a normal number.

Therefore we need to switch to using 'ashr', which does consider the
sign bit:

The 'ashr' instruction (arithmetic shift right) returns the first
operand shifted to the right a specified number of bits with sign
extension.

This change however, exposed another issue. In the tuple handling we
checked whether a fixnum was not too big for the tuple, but not whether
it was positive. This was not a problem before because of the signedness
issue, it would end up as a very big number that was outside the tuple
anyway. Because of switching to 'ashr', we need a better check.

For these better checks we create some helpers, because the check
whether a number of two numbers are positive are needed in a few more
places, such as fixnum_div and fixnum_mod. Using these helpers also
cleans up the code there.