Use regular additions for Fixnum addition and subtraction
Previously we used 63 bit computation with overflows. This however as
slightly different semantics than our Fixnum, because our Fixnum::MIN is
not the same. This would mean that in the JIT a value of Fixnum::MIN - 1
would come out as a fixnum.

Since for multiplication we need these checks anyway, since in LLVM 3.0
i63 with overflow multiplication is broken, we reuse this code for the
addition and subtraction case here. We also immediately promote to
bignum now if it doesn't fit so we don't recompute the value in the
send. This provides a small performance improvement for the case where
Fixnum +/- Fixnum results in a Bignum:

Before:
Fixnum - Fixnum => Bignum
                      6590050.7 (±12.2%) i/s -   32387040 in   5.009288s (cycle=192780)

After:
Fixnum - Fixnum => Bignum
                      6963561.3 (±4.7%) i/s -   34828744 in   5.015573s (cycle=192424)