Don't emit exception checks if inline code doesn't raise
If we know that the inlined code doesn't raise an exception, we don't
need the check. We already did that for the case where the inline used
an uncommon exit for the failure case, but not when the fallback was a
regular send. In that case we always checked for an exception, instead
of just doing this after the send.