Clean Array by using a Reflector (simple mirror)
Array left it's guts exposed to deal with the way that data has to be
exposed in ruby. To clean things up, I've introduce a Reflector class
which is a simple mirror. This can easily access the internal data of an
object. Additionally, rather than using instance_variable_[get|set] the
JIT can make access quite fast because it doesn't have to deal with any
edgecases.