Improve LLVM's ability to reason about data layout
We need to explicitly set the data layout for the LLVM module we use to
compile JIT'ted methods. This allows for example the GVN pass to better
optimize certain cases. One case in this example is:

def method
  1 + 1
  nil
end

Previously LLVM was not able to see through this and completely
eliminate the 1 + 1, with these changes it's able to do so. We also use
the PassManagerBuilder to setup a default set of first passes, similar
to what -O2 means in clang.