Removed RVM from the README.
RVM has repeatedly refused to consider explicit requests not to interfere
with the Rubinius build process. Further, it has repeatedly broken support
for installing Rubinius. There is no reason to promote its use. There are
a number of other tools for installing and using Rubinius, as well as package
mannagers for every major OS distribution. Efforts to make these tools work
well with Rubinius or get packages for major OS distributions is a far better
use of time and effort.

Ultimately, we need a dedicated section in the Rubinius documentation to
explain the various ways in which Rubinius can be installed.