Teach the JIT a better version of check_frozen
This optimizes for the common case of either a non reference or a not
inflated header for a non frozen object. This can be determined by a
check against the object headers, otherwise fall back to the complete
code.

Also introduce defines for the header bits. These can be used instead of
magic constant values inside the JIT.