Setup specific allocators for String
Right now the JIT can't handle super() calls yet. This let to the
allocator for String to never be JIT'ted. For now we use these specific
allocators. In the future we should teach the JIT about super() (at
least for simple cases like this), so we can remove these specifics
then.

Also sets up Character like we setup the other internal VM classes.