Performance improvements for String#gsub
Biggest change here is that originally there was a != check against
undefined. In 1.9 mode, this results in using BasicObject#!, which ended
up in String#==, where the primitive would fail for undefined.

That would trigger the fallback path and try other things such as
coercions before finally resulting in false. This changes the code so it
doesn't unecessary go down that path.