Rename vm_ to state_
In the past, we didn't have the State object, but only VM. This name
change wasn't made in the LLVM parts, which means that even though in
the JIT we used a vm_ object, it was actually the C++ State object in
the vm.

This was still confusing sometimes, so changed the naming to be
consistent with other parts of the vm.