Ensure we add LLVM files and paths last
When using a system LLVM, it might add system wide paths to the include
path. This is fine, but we want to be sure that we add paths to our
vendored libraries before that. Otherwise we might get a conflict when
the system LLVM path also has headers for a vendored library.

Case in point is a FreeBSD system that had libffi in the same paths as
the system LLVM. This caused Rubinius to try to link to this older
version of libffi without proper support for what Rubinius needs. This
caused failures, but only when using system LLVM.