Fix memory corruption problem with Fiber finalization
The problem is that here ->die() is called on the Fiber, which resets
the thread is is associated too. If then the fiber data is deleted, it
will not remove itself from the fiber stacks list of the thread it
belonged to.

This means that when the FiberStacks are cleaned up for this thread, the
fiber data that has been deallocated is still in the list. That leads to
invalid memory writes in vm/fiber_stack.cpp:72 when it calls ->die()
again on the fiber data.

This is the cause of segfaults on Linux when all finalizers are enabled
when Rubinius shuts down. This fix allows for enabling finalizers
properly.