Make debugging easier for Baker GC space
This makes two changes to make some debugging easier. First it marks a
certain as off limits so Valgrind can find errors if objects are used
that still refer to this old space.

The other is that we can easily set a flag to fill the unused space with
all 1's so any usage will very likely result in a segfault and makes
problems appear sooner rather than later.