Always push finalization information on the front
There might be multiple objects that need finalization that depend on
each other. The only really sensible case is an object that is created
later depends on an object created earlier. If both objects needs
finalization, the latest created object should be finalized first.

Changing this ordering here makes sure newest objects are finalized
first, so we work with a LIFO here instead of FIFO.