Reset method cache after defining a method through the C-API
This was problematic with the FFI extension. Problem was that because of
not resetting the method cache, if a function was redefined, Rubinius
would keep using the old cached definition. When that old definition was
cleaned up, since nothing in Ruby land referred to it anymore, it could
crash because of memory already been freed and reused.