Run all finalizers through the scheduling
This is another case where dependencies between finalizers can be
tricky. Data_Wrap_Struct uses a C finalizer for the free function if
that is requested. That means that this data gets cleaned up before any
other finalizers could run that depend on this data.

By treating all finalizers equally here, we can make sure that the
ordering of defining finalizers is also the reverse of when they are
executed.