Prevent Data from being freed twice and marking when freed
There are situations where something happens to hold onto finalized
data, for example in the finalizer itself. Whatever does this or however
this happens, we should prevent double marking and free'ing here as a
safety net.

These objects usually get cleanup up then in the next cycle, it could
for example be that finalization was in progress at that time for those
objects when doing the GC.