Run both finalizers if present
We should run both finalizers if they are set. For example Data objects
with a Ruby finalizer, will have both set. We also need to run both, so
we properly clean up the object. Because of this we should first run the
Ruby finalizer, since the C finalizer might clean up memory etc.