Make interrupt checks in the JIT cheaper
We use the new VMJIT structure to read whether we have an interrupt
scheduled for this thread and check the status of whether we have to GC
from shared state.

Improves performance of for example Integer#gcd, because that checks
interrupts in the loop.

Before:
=== ./bin/rbx ===
Fixnum gcd Fixnum => Fixnum
                      3232489.4 (±3.7%) i/s -   16153410 in   5.004410s

After:
=== bin/rbx ===
Fixnum gcd Fixnum => Fixnum
                      3846171.8 (±2.4%) i/s -   19263285 in   5.011503s