Reworked finalizers.
The objects to be finalized are now grouped into "generations." All the
objects in a generation are kept alive until all the finalizers are run.

The Ruby finalizers are run first, then the C finalizers, and finally, the
objects are "released", which primarily involves cleaning up handle-related
aspects.

The Ruby finalizers are run first so that any objects that are being finalized
can access other objects that were alive at the same time they were alive.

Once all the objects in a generation have been processed, the list of objects
is dropped. Since the objects are kept alive by the list they are a member of,
all objects in a list that is dropped before the next GC cycle will be
collected in that cycle.