Dirkjan found another bug!
Cleaning up inflated headers and handles depends on objects being marked in
the mature GC. If an object needing finalization is young, it will be moved to
a dead list when the young GC runs. However, if a mature GC follows in the
same cycle, the object is not iterated in the dead list. Consequently, if the
object has a handle or inflated header, it is not marked and those are
deallocated. Then all hell breaks loose and unicorns steal baby kittens to use
as gummy bears.