Fix issues in rbx_string_build JIT helper
There were two issues here. First of all the fix from
ff1bdc0036a346d737d88f80cd06f7883eba3e40 wasn't ported to the JIT code.
The other problem was that in the instructions implementation we use the
stack to pass arguments.

Because the stack grows down, we iterate from the highest index to the
lowest. In the JIT however, we iterate through them starting with offset
0. This means that in the JIT code we can ignore the element with index
0, not the one with count - 1.