Setup objects as dirty where possible
There are quite a few places where we end up initializing an object
completely. In these cases we can allocate them as dirty, because all
instance variables will be set anyway. This prevents setting these
variables twice, first with nil and then with the actualy value.

These are the places that showed up in CI runs as most used.