Move constants for C-Api constants into separate header file
This makes it easier to remove the static global map we now keep and
track this in the shared state. This also moves copied / shared header
files between 1.8 and 1.9 C-Api into a separate header and copies them
for the different modes.

The mri_oop.h file was created staticly from the configure script. If
it's static anyway, we might as well copy it from a file instead of
having it all inline in configure.