Move C-API constants into shared state
This avoids having global data for C-API constants. It also makes it
easier to wrap changes to these structures with a lock, which causes
issues like #2138. This change therefore also adds that lock so the list
of globals isn't modified concurrently.

This also would improve future possibility of embedded, as long as
people don't stuff things into global variables. The latter case is
already problematic anyway for extensions, so this change doesn't make
this problem worse and improves by not having the global state.