Setup Ruby side object for ThreadState
For code that heavily uses ensure blocks, creating these thread state
objects had quite a big overhead. We were abusing the Rubinius::Internal
exception for this behavior and spend a lot of time for that code in
symbol lookup for the instance variables.

This solution removes the need for any symbol lookup and this also helps
in concurrent scenarios since symbol lookup needs a lock around the
symbol table.